backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pymongo==4.12.1
dnspython==2.7.0
pydantic==2.11.3
pydantic-settings==2.9.1
sse-starlette==2.3.3
python-dotenv==1.1.0
httpx==0.28.1
PyJWT==2.9.0
